pyproject.toml
[tool.poetry]
name = "danbiedu"
version = "0.1.0"
description = ""
authors = ["sgaeng <ksge1124@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
djangorestframework = "^3.14.0"
pyjwt = "^2.6.0"
django-environ = "^0.10.0"
django-cors-headers = "^3.14.0"
psycopg2-binary = "^2.9.6"
drf-yasg = "^1.21.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"